Migrate worker page to new js workflow
