+++ presto-main/src/main/resources/webapp/worker.html
@@ -105,8 +105,7 @@
 window.setInterval(updateClusterInfo, 1000)
 </script>
 
-<script src="vendor/react/browser.min.js"></script>
-<script type="text/babel" src="assets/worker.js"></script>
+<script type="text/javascript" src="dist/worker.js"></script>
 
 <!-- Fonts -->
 <link href="https://fonts.googleapis.com/css?family=Roboto:400,500,700" rel="stylesheet">
